nuevo back, falta testear

+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/categoriaAI/CategoriaAIRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoriaAIRepository extends JpaRepository<CategoriaAIEntity, Long>{
+public interface CategoriaAIRepository extends JpaRepository<CategoriaAIEntity, Integer>{
 
 }
