+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/configuracionGeneral/ConfiguracionGeneralRepository.java
@@ -4,4 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConfiguracionGeneralRepository extends JpaRepository<ConfiguracionGeneralEntity, Long>{}
+public interface ConfiguracionGeneralRepository extends JpaRepository<ConfiguracionGeneralEntity, Integer>{
+	
+}
