+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/detallePago/DetallePagoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetallePagoRepository extends JpaRepository<DetallePagoEntity, Long>{}
+public interface DetallePagoRepository extends JpaRepository<DetallePagoEntity, Integer>{}
