+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/detallePedido/DetallePedidoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetallePedidoRepository extends JpaRepository<DetallePedidoEntity, Long>{}
+public interface DetallePedidoRepository extends JpaRepository<DetallePedidoEntity, Integer>{}
