+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/detalleReceta/RecetaDetalleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecetaDetalleRepository extends JpaRepository<RecetaDetalleEntity, Long>{
+public interface RecetaDetalleRepository extends JpaRepository<RecetaDetalleEntity, Integer>{
 
 }
