+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/factura/FacturaRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FacturaRepository extends JpaRepository<FacturaEntity, Long>{}
+public interface FacturaRepository extends JpaRepository<FacturaEntity, Integer>{}
