+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/formaPago/FormaPagoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FormaPagoRepository extends JpaRepository<FormaPagoEntity, Long>{}
+public interface FormaPagoRepository extends JpaRepository<FormaPagoEntity, Integer>{}
