+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/receta/RecetaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecetaRepository extends JpaRepository<RecetaEntity, Long>{
+public interface RecetaRepository extends JpaRepository<RecetaEntity, Integer>{
 
 }
