+++ backend/buenGusto-final/src/main/java/com/utn/app/buenGusto/rol/RolRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RolRepository extends JpaRepository<RolEntity, Long>{}
+public interface RolRepository extends JpaRepository<RolEntity, Integer>{}
